+++ CHANGELOG.md
@@ -14,13 +14,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Family::get_or_create_owned` can access a metric in a labeled family. This
    method avoids the risk of runtime deadlocks at the expense of creating an
    owned type. See [PR 244].
- 
+
 [PR 244]: https://github.com/prometheus/client_rust/pull/244
 [PR 257]: https://github.com/prometheus/client_rust/pull/257
 
 ### Changed
 
 - `EncodeLabelSet::encode()` now accepts a mutable reference to its encoder parameter.
+- Emit better compilation error message when deriving marcos `EncodeLabelSet` and `EncodeLabelValue`.
+  See [PR 267].
+
+### Fixed
+
+- Fixed an issue where the derive marcos `EncodeLabelSet` and `EncodeLabelValue` didn't work
+  when the `struct` has generic parameters (like `'a`).
+  See [PR 265].
+
+- Fixed an issue where the derive macro `EncodeLabelSet` didn't work
+  when the `struct` has generic parameters (like `'a`).
+  See [PR 267].
+
+[PR 265]: https://github.com/prometheus/client_rust/pull/265
+[PR 267]: https://github.com/prometheus/client_rust/pull/266
 
 ## [0.23.1]